Update launch file for 3 cameras (#74)
* Update launch file for 3 cameras

* Rename to eve

* Fix Linting
