--- src/perception/camera_object_detection/config/sim_config.yaml
@@ -1,7 +1,0 @@
-camera_object_detection_node:
-  ros__parameters:
-    camera_topic: /camera/right/image_color
-    publish_vis_topic: /annotated_img
-    publish_obstacle_topic: /obstacles
-    model_path: /perception_models/yolov8s.pt
-    image_size: 480
